replayer: Filter out CLion Doctest arguments
Our fuzz tests pretend to be Doctest tests and thus need to ignore the
arguments CLion's test framework integration passes to such tests.

With this commit and #19, play buttons are on fuzz tests in CLion
successfully trigger a regression test run on the seed corpus, with
command output shown in a console. The test status is currently shown
as "testing framework exited unexpectedly" after the run, regardless of
whether the test passed or not. This will be improved in a follow-up PR.