replayer: Print summary and follow-up commands on exit
Regardless of whether the replayer exits normally or abnormally, print a
summary of the result of the regression test run and a follow-up command:
If the test passed, inform the user that no actual fuzzing has been
performed and print the cifuzz run command required to start an actual
fuzzing run.
If the test failed, print a command with which the user can debug the
issue. This is Linux-only for now until we implement the debug
subcommand.