create: Simplify flags
Instead of having one flag to specify the output directory and one to
specify the filename, there is now only one flag to specify the output
path.

This also simplifies the code, allowing to get rid of the "pkg/storage"
package.

Note that the behavior changed a bit, in that output directories are not
created automatically anymore and creation of the fuzz target stub fails
if the output directory doesn't exist yet - which is what I would expect
from a command which creates a file.