bundle: Include external non-system libraries in the archive
This covers cases such as a CMake target that depends on a library
provided by Conan. Such a library is not contained in the current CMake
project's build directory, but still created as part of the build and
thus not a good candidate for inclusion into the static Docker image
used as the runtime environment.