run: Support specifying command options via config file
This uses viper to read the config file and set the command options
according to the settings in the config file. The commit also merges
the old project config into the command options struct, because the only
setting of the project config, the build system setting, can just as
well be a command option (and should also be configurable via a
command-line flag - that will be done in a follow-up commit).

Using viper allows us to handle configuration with very little
self-maintained code. What motivated me to make this change is the
viper.SetDefault() function, which allows us to set a default value
which is applied when neither the config file setting nor the flag is set,
a functionality that we would have reimplement otherwise.