cmake: Build the replayer with the fuzz test
This gets rid of a big bag of hacks and cleanly ensures that the
replayer is built with the same flags as the fuzz test, which enables
e.g. cross-compilation scenarios.

Additionally, it saves us from committing to either the legacy or the
modern signature of target_link_library, both of which can result in
warnings or errors depending on the user's CMake version.