all: Let the caller handle ExecErrors
The executil.HandleExecError function used to log error messages before
the caller could decide what to do with the error. We now wrap the error
in a cmdutils.ExecError and let the caller decide whether to print the
error message or not.

This avoids printing two errors when Ctrl+C is used to interrupt a run:

    ❌ minijail0: exit status 72
    ❌ terminated by signal 2 (interrupt)

and just prints the second error instead (as was intended).