Update README.md: Reordered installation instructions and added some explanations
The prerequisite installation instructions were split in two parts. I reordered the instructions so that first all dependency installations are done and then cifuzz is installed.
Also some minor changes were done to the "create your first fuzz test" to make it clear that the instructions printed by the CLI should also be followed and that additional directives need to be added to CMakeLists.txt to link a fuzz test with the software under test.