parser test: Fix stack traces created on Windows
The logs used in the tests contained absolute Unix paths, which caused
the validateSourceFile function to treat them as valid source file
paths. By using Windows paths when running on Windows, the
validateSourceFile function now correctly treats them as invalid source
file paths (because the files are not below the project directory).