installer: Fix installation of bash completion script on macOS
On macOS, we used to always install the bash completion script to
"$(brew --prefix)/etc/bash_completion.d/cifuzz", which was not always
sourced, causing the command completion to not work after successfuly
installation.

We now tell the user to source our installed command completion script
in their ~/.bash_profile.